Create Artman class
